v1.1.0

+++ index.js
@@ -1,11 +1,19 @@
 const fs = require("node:fs");
 const path = require("node:path");
+const nconf = require("nconf");
+
+const configPath = path.join(__dirname, "src/configs/config.json");
+nconf.file("default", configPath);
+
+let root = nconf.get("root");
+
 const {
 	Client,
 	Collection,
 	Events,
 	GatewayIntentBits,
 	ChannelType,
+	ActivityType,
 } = require("discord.js");
 require("dotenv").config();
 const wait = require("node:timers/promises").setTimeout;
@@ -23,7 +31,6 @@ async function countLines(input) {
 	return lineCount;
 }
 const client = new Client({ intents: [GatewayIntentBits.Guilds] });
-var domain = "";
 client.commands = new Collection();
 const commandsPath = path.join(__dirname, "src/commands");
 const commandFiles = fs
@@ -38,32 +45,35 @@ for (const file of commandFiles) {
 
 client.once(Events.ClientReady, async () => {
 	console.log("Ready!");
+
+	client.user.setActivity("better side..", { type: ActivityType.Watching });
+
 	guild = client.guilds.cache.get(process.env.GUILD_ID);
 	if (
 		!guild.channels.cache.find(
 			(channel) =>
 				channel.type == ChannelType.GuildCategory &&
-				channel.name == process.env.CATEGORY_NAME
+				channel.name == nconf.get("categoryName")
 		)
 	) {
 		await guild.channels.create({
-			name: process.env.CATEGORY_NAME,
+			name: nconf.get("categoryName"),
 			type: ChannelType.GuildCategory,
 		});
 	}
 	let category = await guild.channels.cache.find(
 		(channel) =>
 			channel.type == ChannelType.GuildCategory &&
-			channel.name == process.env.CATEGORY_NAME
+			channel.name == nconf.get("categoryName")
 	);
 	if (
 		!guild.channels.cache.find(
-			(c) => c.name.toLowerCase() === process.env.CHANNEL_NAME
+			(c) => c.name.toLowerCase() === nconf.get("channelName")
 		)
 	) {
 		guild.channels
 			.create({
-				name: process.env.CHANNEL_NAME,
+				name: nconf.get("channelName"),
 				type: ChannelType.GuildText,
 			})
 			.then((channel) => {
@@ -73,47 +83,110 @@ client.once(Events.ClientReady, async () => {
 });
 
 client.on(Events.InteractionCreate, async (interaction) => {
-	if (
-		client.channels.cache.get(interaction.channelId).name ==
-		process.env.CHANNEL_NAME
-	) {
-		if (!interaction.isChatInputCommand()) return;
+	if (!interaction.isChatInputCommand()) return;
+	nconf.file("default", configPath);
+	let users = Object.assign([], nconf.get("users"));
 
-		if (interaction.commandName === "domain") {
-			let subCommand = interaction.options.getSubcommand();
-			if (subCommand == "manage") {
-				domain = interaction.options.getString("name");
-				require("./src/actions/domainOptions").manage(interaction, domain);
-			}
-			if (subCommand == "list") {
-				require("./src/actions/domainList").list(interaction);
-			}
-			if (subCommand == "certificates") {
-				require("./src/actions/domainList").certificates(interaction);
-			}
-		}
-	} else {
+	if (root != interaction.user.tag && !users.includes(interaction.user.tag)) {
+		interaction.reply({
+			content: `You are not authorized!`,
+			ephemeral: true,
+		});
+		return;
+	}
+	if (
+		client.channels.cache.get(interaction.channelId).name !=
+		nconf.get("channelName")
+	) {
 		interaction.reply({
 			content: `Wrong channel. Correct: 👉 <#${
 				guild.channels.cache.find(
-					(c) => c.name.toLowerCase() === process.env.CHANNEL_NAME
+					(c) => c.name.toLowerCase() === nconf.get("channelName")
 				).id
 			}>`,
 			ephemeral: true,
 		});
+		return;
+	}
+
+	if (interaction.commandName === "domain") {
+		let subCommand = interaction.options.getSubcommand();
+		if (subCommand == "manage") {
+			domain = interaction.options.getString("name");
+			require("./src/actions/domainOptions").manage(interaction, domain);
+		}
+		if (subCommand == "list") {
+			require("./src/actions/domainList").list(interaction);
+		}
+		if (subCommand == "certificates") {
+			require("./src/actions/domainList").certificates(interaction);
+		}
+	}
+	if (interaction.commandName === "webserver") {
+		let subCommand = interaction.options.getSubcommand();
+		let subCommandGrup = interaction.options.getSubcommandGroup();
+
+		if (subCommand == "off") {
+			require("./src/actions/serverManage").off(interaction);
+		}
+		if (subCommand == "on") {
+			require("./src/actions/serverManage").on(interaction);
+		}
+
+		if (interaction.user.tag == root) {
+			if (subCommandGrup == "message-visibility") {
+				if (subCommand == "hiden") {
+					require("./src/actions/serverManage").hide(interaction);
+				}
+				if (subCommand == "visible") {
+					require("./src/actions/serverManage").show(interaction);
+				}
+			}
+			if (subCommandGrup == "users") {
+				if (subCommand == "add") {
+					require("./src/actions/serverManage").userAdd(interaction);
+				}
+				if (subCommand == "remove") {
+					require("./src/actions/serverManage").userRemove(interaction);
+				}
+				if (subCommand == "list") {
+					require("./src/actions/serverManage").usersList(interaction);
+				}
+			}
+		} else {
+			interaction.reply({
+				content: `You are not permission!`,
+				ephemeral: true,
+			});
+		}
 	}
 });
-
 client.on(Events.InteractionCreate, async (interaction) => {
 	if (!interaction.isButton()) return;
+	nconf.file("default", configPath);
+	let users = Object.assign([], nconf.get("users"));
+	if (root != interaction.user.tag && !users.includes(interaction.user.tag)) {
+		interaction.reply({
+			content: `You are not authorized!`,
+			ephemeral: true,
+		});
+		return;
+	}
+	let domain = interaction.message.embeds[0].data.fields[0].value;
+
 	if (
 		interaction.customId === "add" ||
 		interaction.customId === "remove" ||
 		interaction.customId === "update" ||
 		interaction.customId === "changeProtocol"
 	) {
-		await interaction.message.delete();
-		await interaction.deferReply();
+		if (!nconf.get("messageVisibility")) {
+			await interaction.message.delete();
+
+			await interaction.deferReply({
+				ephemeral: nconf.get("messageVisibility"),
+			});
+		}
 	}
 	if (interaction.customId === "add") {
 		require("./src/actions/domainAdd").protocol(interaction, domain);
@@ -128,7 +201,24 @@ client.on(Events.InteractionCreate, async (interaction) => {
 		require("./src/actions/domainRemove").remove(interaction, domain);
 	}
 	if (interaction.customId === "cancel") {
-		await interaction.message.delete();
+		if (!nconf.get("messageVisibility")) {
+			await interaction.message.delete();
+		} else {
+			let enabled = {
+				color: 0xff0000,
+				title: "canceled!",
+				footer: {
+					text: "made by ~ kacpep.dev",
+					icon_url: "https://i.imgur.com/M0uWxCA.png",
+				},
+			};
+			await interaction.update({
+				content: "",
+				embeds: [enabled],
+				components: [],
+				ephemeral: nconf.get("messageVisibility"),
+			});
+		}
 	}
 	if (interaction.customId === "changeProtocol") {
 		require("./src/actions/domainUpdate").selectChangeProtocol(
@@ -146,10 +236,26 @@ client.on(Events.InteractionCreate, async (interaction) => {
 
 client.on(Events.InteractionCreate, async (interaction) => {
 	if (!interaction.isStringSelectMenu()) return;
-	if (interaction.customId === "selectProtocol") {
-		await interaction.message.delete();
+	nconf.file("default", configPath);
+	let users = Object.assign([], nconf.get("users"));
 
-		await interaction.deferReply();
+	if (root != interaction.user.tag && !users.includes(interaction.user.tag)) {
+		interaction.reply({
+			content: `You are not authorized!`,
+			ephemeral: true,
+		});
+		return;
+	}
+
+	let domain = interaction.message.embeds[0].data.fields[0].value;
+
+	if (interaction.customId === "selectProtocol") {
+		if (!nconf.get("messageVisibility")) {
+			await interaction.message.delete();
+			await interaction.deferReply({
+				ephemeral: nconf.get("messageVisibility"),
+			});
+		}
 
 		if (interaction.values == 443) {
 			require("./src/actions/domainAdd").addHTTPS(interaction, domain);
@@ -158,7 +264,7 @@ client.on(Events.InteractionCreate, async (interaction) => {
 		}
 	}
 	if (interaction.customId === "selectChangeProtocol") {
-		await interaction.message.delete();
+		if (!nconf.get("messageVisibility")) await interaction.message.delete();
 
 		if (interaction.values == 443) {
 			if (
@@ -166,7 +272,10 @@ client.on(Events.InteractionCreate, async (interaction) => {
 					fs.createReadStream(path.join("/etc/nginx/sites-available", domain))
 				)) <= 10
 			) {
-				await interaction.deferReply({ ephemeral: false });
+				if (!nconf.get("messageVisibility"))
+					await interaction.deferReply({
+						ephemeral: nconf.get("messageVisibility"),
+					});
 
 				require("./src/actions/domainUpdate").changeToHTTPS(
 					interaction,
@@ -174,7 +283,6 @@ client.on(Events.InteractionCreate, async (interaction) => {
 				);
 			} else {
 				await interaction.deferReply({ ephemeral: true });
-
 				await interaction.editReply({
 					content: `You cannot change the protocol to the same!`,
 					ephemeral: true,
@@ -186,7 +294,10 @@ client.on(Events.InteractionCreate, async (interaction) => {
 					fs.createReadStream(path.join("/etc/nginx/sites-available", domain))
 				)) <= 10
 			) {
-				await interaction.deferReply({ ephemeral: false });
+				if (!nconf.get("messageVisibility"))
+					await interaction.deferReply({
+						ephemeral: nconf.get("messageVisibility"),
+					});
 
 				require("./src/actions/domainUpdate").changeToHTTP(interaction, domain);
 			} else {
@@ -203,7 +314,19 @@ client.on(Events.InteractionCreate, async (interaction) => {
 
 client.on(Events.InteractionCreate, async (interaction) => {
 	if (!interaction.isModalSubmit()) return;
-	await interaction.message.delete();
+	nconf.file("default", configPath);
+	let users = Object.assign([], nconf.get("users"));
+
+	if (root != interaction.user.tag && !users.includes(interaction.user.tag)) {
+		interaction.reply({
+			content: `You are not authorized!`,
+			ephemeral: true,
+		});
+		return;
+	}
+
+	let domain = interaction.message.embeds[0].data.fields[0].value;
+	if (!nconf.get("messageVisibility")) await interaction.message.delete();
 
 	if (interaction.customId === "modalPortForwarding") {
 		let port = interaction.fields.getTextInputValue("newPort");
@@ -222,3 +345,10 @@ client.on(Events.InteractionCreate, async (interaction) => {
 });
 
 client.login(process.env.DISCORD_TOKEN);
+
+process.on("uncaughtException", (error) => {
+	console.log("-----handler-------");
+	console.log(error);
+	console.log("-----handler-------");
+	process.exit(1);
+});
